homework_02/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==2.4.0
selenium==3.141.0
webdriver-manager==3.5.1
xdist==0.0.1
